server/requirements.txt
flask-login
flask-sqlalchemy
flask-wtf
google-api-python-client
google-auth-oauthlib
pip-chill
python-dotenv
